refactor: 기존 NoiseMap 구조 리팩토링
